clean up client / fix fe
